납품전 버그수정

+++ admin/api/pcs/ordersheet/dsn/wkla/delete.php
@@ -13,15 +13,24 @@
  | 
  +=============================================================================
 */
+
 $sel_idx        = $_POST['sel_idx'];
 
-$sql = 	'
+$product_cnt = null;
+$product_cnt = $db->count('ORDERSHEET_MST', "WKLA_IDX = ".$sel_idx );
+
+if($product_cnt == 0){
+    $sql = 	'
         DELETE FROM
-            dev.WKLA_INFO
+            WKLA_INFO
         WHERE
             IDX = '.$sel_idx.'
-';
+    ';
 
-$db->query($sql);
-
+    $db->query($sql);
+}
+else if($product_cnt > 1){
+    $code = 300;
+    $msg = '이미 사용중인 상품이 있는 W/K/L/A은 삭제할 수 없습니다.';
+}
 ?>
