+++ admin/api/pcs/ordersheet/md/line/delete.php
@@ -15,13 +15,22 @@
 */
 $sel_idx        = $_POST['sel_idx'];
 
-$sql = 	'
+$product_cnt = null;
+$product_cnt = $db->count('ORDERSHEET_MST', "LINE_IDX = ".$sel_idx );
+
+if($product_cnt == 0){
+    $sql = 	'
         DELETE FROM
-            dev.LINE_INFO
+            LINE_INFO
         WHERE
             IDX = '.$sel_idx.'
-';
+    ';
 
-$db->query($sql);
+    $db->query($sql);
+}
+else if($product_cnt > 1){
+    $code = 300;
+    $msg = '이미 사용중인 상품이 있는 라인은 삭제할 수 없습니다.';
+}
 
 ?>
