+++ admin/api/pcs/ordersheet/td/sub_material/delete.php
@@ -13,15 +13,34 @@
  | 
  +=============================================================================
 */
-$sel_idx        = $_POST['sel_idx'];
 
-$sql = 	'
-        DELETE FROM
-            dev.SUB_MATERIAL_INFO
-        WHERE
-            IDX = '.$sel_idx.'
-';
+$sub_material_idx		= $_POST['sub_material_idx'];
 
-$db->query($sql);
+$product_cnt = $db->count(
+	"ORDERSHEET_MST OM",
+	"OM.IDX IN (SELECT DISTINCT S_SM.ORDERSHEET_IDX FROM SUB_MATERIAL_MAPPING S_SM WHERE S_SM.SUB_MATERIAL_IDX = ".$sub_material_idx.") AND OM.DEL_FLG = FALSE"
+);
+
+if ($product_cnt > 0) {
+	$json_result['code'] = 300;
+	$json_result['msg'] = '현재 사용중인 부자재 정보는 삭제할 수 없습니다.';
+} else {
+	$delete_sub_material_sql = "
+		DELETE FROM
+			SUB_MATERIAL_INFO
+		WHERE
+			IDX = ".$sub_material_idx."
+	";
+	
+	$db->query($delete_sub_material_sql);
+
+	$delete_sub_material_img_sql = "
+		DELETE FROM
+			SUB_MATERIAL_IMAGE
+		WHERE
+			SUB_MATERIAL_IDX = ".$sub_material_idx."
+	";
+	$db->query($delete_sub_material_img_sql);
+}
 
 ?>
